requirements.txt
langchain>=0.1.16
langchain-community>=0.0.34
gradio>=4.24.0
chromadb>=0.4.24
sentence-transformers>=2.2.2
python-dotenv>=1.0.1
ibm-watsonx-ai>=0.1.2
PyPDF2>=3.0.1
